webapi/webapi-dlna-xwalk-tests/dlna/MediaServer_attributes_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang, Hongjuan <hongjuanx.wang@intel.com>

-->

<meta charset="utf-8">
<title>DLNA Test: MediaServer - attributes basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://01org.github.io/cloud-dleyna/mediaserver.html#mediaserver-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var msga, mse;
  setup(function () {
    msga = navigator.mediaServer;
    msga.scanNetwork();
  });

  var t = async_test("the onserverfound event can be invoked");

  on_event(msga, "serverfound", function (event) {
    mse = event.server;

    test(function () {
      assert_equals(mse.friendlyName, "iMediaExport", "the MediaServer.friendlyName value");
    }, "Check if the MediaServer.friendlyName value is user-friendly name for the device");

    test(function () {
      assert_equals(mse.manufacturer, "Rygel Developers.", "the MediaServer.manufacturer value");
    }, "Check if the MediaServer.manufacturer value is a string identifying the manufacturer of the server");

    test(function () {
      assert_equals(mse.manufacturerURL, "http://live.gnome.org/Rygel", "the MediaServer.manufacturerURL value");
    }, "Check if the MediaServer.manufacturerURL value is a URL pointing to the manufacturer's web site");

    test(function () {
      assert_equals(mse.modelName, "Rygel", "the MediaServer.modelName value");
    }, "Check if the MediaServer.modelName value is model name of the server");

    test(function () {
      assert_equals(mse.modelNumber, "0.21.1.1-a97e", "the MediaServer.modelNumber value");
    }, "Check if the MediaServer.modelNumber value is the server's version number");

    test(function () {
      assert_equals(mse.serialNumber, "0000001", "the MediaServer.serialNumber value");
    }, "Check if the MediaServer.serialNumber value is the server's serial number");

    test(function () {
      assert_equals(mse.iconURL.substr(mse.iconURL.lastIndexOf("/")+1), "MediaExport-48x48x24.jpg",
      "the MediaServer.iconURL value");
    }, "Check if the MediaServer.iconURL value is the URL pointing to an icon that graphically identifies the server");

    test(function () {
      assert_equals(mse.deviceType, "urn:schemas-upnp-org:device:MediaServer:3", "the MediaServer.deviceType value");
    }, "Check if the MediaServer.deviceType value is the UPnP type of the device");

    t.done();
  });
</script>
